+++ src/main.ts
@@ -3,6 +3,6 @@ import '@/styles/global.css'
 import '@/styles/tailwind.css'
 import App from '@/App.vue'
 import router from '@/router'
-import store from '@/store'
+import store from 'src/stores'
 
 createApp(App).use(router).use(store).mount('#app')
